% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot.R
\name{cv_plot}
\alias{cv_plot}
\title{cv_plot}
\usage{
cv_plot(df, ni, B, nb, x_label, y_label)
}
\arguments{
\item{df}{A data frame containing two columns: Factors and X. The X column represents the vector of bootstrapped values for analysis.}

\item{ni}{An integer represents the starting value of the cumulative series.}

\item{B}{An integer that determines the number of resamplings.}

\item{nb}{An integer indicates the increasing number of the cumulative series.}

\item{x_label}{A character vector with the label for the X-axis.}

\item{y_label}{A character vector with the label for the y-axis.}
}
\value{
A plot of \emph{CV} results against the cumulative number of iterations for each model.
}
\description{
Plot of cumulative coefficient of variation.
}
\details{
The efficiency of the resamples in the bootstrap simulations was analyzed
graphically using the number of resamples \emph{versus} the coefficient of variation
\emph{CV}. The \emph{CV} was calculated for the cumulative resamples of each model,
with the cumulative values increasing after every \emph{ni} data point.

cv=sd(X)/mean(X)*100

where \emph{x} represents the vector of bootstrapped values and \emph{sd} denotes
the standard deviation.

A data frame named \code{df} is required to use this function. It should include a
Factor column with its factor levels and resampled values in the vector \emph{X},
where \emph{X} is a capital letter.

The labels for the X and Y axes were defined in \code{x_label} and \code{y_label}, respectively.
}
\examples{
\dontrun{
set.seed(16) # Set seed for reproducibility

n <- 1000  # Sample size
mean_1 <- 10  # Mean of the distribution
sd_1 <- 3  # Standard deviation of the distribution
sample_1 <- rnorm(n, mean_1, sd_1)

mean_2 <- 15  # Mean of the distribution
sd_2 <- 2  # Standard deviation of the distribution
sample_2 <- rnorm(n, mean_2, sd_2)

X <- rbind(sample_1,sample_2) # bind resampled values in one vector
Factor <- as.factor(rep(c("sample_1","sample_2"), each=n))

df <- data.frame(Factor,X= as.numeric(X))

x_label <- "Cummulative iterations"
y_label <- "CV\%"

ni <- 50
B <- n
nb <- 50

cv_plot <- cv_out(df,ni,B,nb, x_label, y_label)
cv_plot
 }

}
