% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_mature.R
\name{freq_mature}
\alias{freq_mature}
\title{freq_mature}
\source{
\url{https://github.com/Macrurido/Repbio.git}
}
\usage{
freq_mature(x_total, x_active, Imin, Imax, bin)
}
\arguments{
\item{x_total}{A numeric vector of the length of total of organisms in the sample}

\item{x_active}{A numeric vector of the length of reproductive active organisms}

\item{Imin}{A numeric value of the shortest organism.}

\item{Imax}{A numeric value of the longest organism.}

\item{bin}{A numeric value that represents the amplitude of the class interval.}
}
\value{
A data frame with the frequencies and proportions of a binomial
variable expressed as a fraction of n; for example, the proportion
of active organisms to the total number of organisms in the ith class interval.
}
\description{
The function calculates the absolute frequency of the total number of
organisms \strong{x_total} and the organisms active during reproduction by class
interval \strong{x_active}.
}
\details{
It provides a data frame contains four variables (columns): the class mark \code{CM},
the absolute frequency of the total number of organisms \code{ni}, the  number of
organisms active during reproduction \code{mi}, and the proportions of reproductive
active organisms in each length class.

This function utilizes the \code{fn_freq()} function from the \code{Repbio} package.
}
\examples{
\dontrun{
freq_mature <- freq_mature(x_total, x_active, Imin, Imax, bin)
}

}
\seealso{
fn_freq() from the \code{Repbio} package
}
