% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaM.R
\name{GaM}
\alias{GaM}
\title{GaM}
\usage{
GaM(prm, x)
}
\arguments{
\item{prm}{a list that stores the gamma model's parameter values b1 and b2.}

\item{x}{is a numeric vector of the ith class interval.}
}
\value{
a numerical vector with the estimated values.
}
\description{
The Gamma model function uses the \code{pgamma()} function from the \code{stats}
package to find the value of the cumulative density function of a gamma
distribution with certain shape \strong{b1} and rate parameters \strong{b2}.
}
\details{
\strong{b1} is a numerical value that affects the the functional form of the curve,
while \strong{b2} is a numerical value that influences the width of the interval.
Both are stored in the \code{prm} list.
}
\examples{
\dontrun{
x <- seq(10,30,by=2.5)
b1 <- 26.8
b2 <- 1.44
prm <- c("b1"=b1,"b2"=b2)
y_hat <- GaM(prm, x)
}

}
\seealso{
pgamma() function from the stats package
}
