% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad.R
\name{glad}
\alias{glad}
\title{glad}
\usage{
glad(models, prm, df, x_label, y_label)
}
\arguments{
\item{models}{A list that contains the functions of the models.}

\item{prm}{A list that contains the parameters values for each model.}

\item{df}{A data frame with the x (class mark), y (proportion) and an empty column y_hat filled with NA}

\item{x_label}{A character vector with the label for the X-axis.}

\item{y_label}{A character vector with the label for the y-axis.}
}
\value{
a list with the new parameter values.
}
\description{
The \code{glad()} function stores the graphically determined values of the start
parameters of the models you want to analyze. Each of the models is analyzed
through a for loop; for this, the \code{simodiff::again()} function is nested.
}
\details{
For more information, please refer to the help section of the \code{simodiff::again()}
function.
}
\examples{
\dontrun{
i <- 1
x <- seq(10,30,by=2.5)
y <- y
b1 <- 38
b2 <- 8
prm[[i] <- c("b1"=b1,"b2"=b2)
prm <- prm[[i]]
df <- cbind(x, y)
df <- as.data.frame(df)
x_label <- "Class mark (cm)"
y_label <- "Proportion"
prm <- glad(models, prm, df, x_label, y_label)
}

}
\seealso{
again function from simodiff package.
}
