% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad_in.R
\name{glad_in}
\alias{glad_in}
\title{glad_in}
\usage{
glad_in(i, models, prm, df, x, x_label, y_label)
}
\arguments{
\item{i}{An integer that specifies which model will be analyzed.}

\item{models}{A list that contains the functions of the models.}

\item{prm}{A list that contains the parameters values for each model.}

\item{df}{A data frame with the x (class mark), y (proportion) and an empty column y_hat filled with NA}

\item{x}{A numeric vector of the ith class interval.}

\item{x_label}{A character vector with the label for the X-axis.}

\item{y_label}{A character vector with the label for the y-axis.}
}
\value{
a list with the new parameter values.
}
\description{
An interactive process to visually determine the initial parameters.
}
\details{
The \code{glad_in()} function is a nested function within the \code{glad()} function.
It conducts an interactive routine to determine the initial parameter values
that will be utilized later by the \code{optim()} optimizer for a specific model.

\strong{\emph{NOTE:}} The \code{glad_in()} function must be executed in the console; when
run from RMarkdown, the prompt appears, but the figure does not.

An iterative process is performed to visually estimate the values of the model's
parameters. This procedure can be repeated until the user is satisfied with
the input values of the analyzed model's parameters.

The start values of the parameters are stored in the \code{prm} list, outside of
this function.

The parameters values are adjusted visually by using a plot, aiming to find a
suitable starting value. If the user is satisfied with the values, they are
stored in the list \code{prm[[i]]}; otherwise, the loop is repeated until the user
is content with the approximations.

The plot of the observed and predicted data using a sigmoidal model is
displayed in a separate window. The estimated value \code{y_hat} (solid line),
is calculated based on the chosen model and the respective initial or revised
values of the parameters: \code{y_hat <- models[[i]](prm[[i]],x)}.

The console prompts you with the question: \strong{Are you happy with the model values? 1=Yes, 0= No}
If you choose 1, it means "yes," and if you choose 0, it means "no."
Type the appropriate number and press Enter.

If you are visually satisfied with the adjusted parameters, type 1 to end the
repeat loop.

If you type 0, you will need to provide the new values for parameters b1 and b2.
Enter the new value for b1 and press Enter, then do the same for b2. If you
accidentally press Enter, a warning pop-up will appear, informing you to enter
a numerical value.

After entering the new values for b1 and b2, the loop will repeat, displaying
the graph with the updated parameters and asking you to confirm if you are
satisfied with the new values.

\strong{WARNING:} The \code{glad_in()} function must be executed in the console;
when run from RMarkdown, the prompt appears, but the figure does not.
}
\examples{
\dontrun{
i <- 1

models <- list(LoM= simodiff::LoM)
prm <- list(c("b1"=38,"b2"=8))

x <- seq(16.25,66.25, by=2.5)
y <- c(0.06, 0.00, 0.01, 0.01, 0.08, 0.04, 0.04,
       0.06, 0.58, 0.42, 0.69, 0.80, 1.00, 0.83,
       1.00, 1.00, 1.00, 1.00, 1.00,  NaN, 1.00)

df <- cbind(x, y)
df <- as.data.frame(df)

x_label <- "Class mark (cm)"
y_label <- "Proportion"

prm <- glad(models, prm, df, x_label, y_label)
}

}
\seealso{
Package ggplot2 version 3.5.1

splot function from simodiff package.
}
