% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoM.R
\name{GoM}
\alias{GoM}
\title{GoM}
\usage{
GoM(prm, x)
}
\arguments{
\item{prm}{a list that stores the Gompertz model's parameter values b1 and b2.}

\item{x}{is a numeric vector of the ith class interval.}
}
\value{
a numerical vector with the estimated values.
}
\description{
The Gompertz model function was implemented to find the value of the
cumulative density function of a Gompertz distribution with certain shape \strong{b1}
and rate parameters \strong{b2}.
}
\details{
pi=exp(-exp(-b2*(x-b1)))

Where:

pi represents a binomial variable expressed as a
fraction of n; for example, the proportion of active organisms to
the total number of organisms in the ith class interval.

b1 is the point of inflection of the curve.

b2  is related directly to the width of the interval.

exp  is the base of the natural logarithm (2.71828...)
}
\examples{
\dontrun{
 x <- seq(10,30,by=2.5)
b1 <- 35
b2 <- 0.11
prm <- c("b1"=b1,"b2"=b2)
y_hat <- GoM(prm, x)
}

}
