% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiM.R
\name{HiM}
\alias{HiM}
\title{HiM}
\usage{
HiM(prm, x)
}
\arguments{
\item{prm}{a list that stores the Hill model's parameter values b1 and b2.}

\item{x}{is a numeric vector of the ith class interval.}
}
\value{
a numerical vector with the estimated values.
}
\description{
The Hill model function was implemented to find the value of the
cumulative density function of a Hill distribution with certain shape \strong{b1}
and rate parameters \strong{b2}.
}
\details{
pi=1/(1+(b1/x)**b2)

Where:

pi represents a binomial variable expressed as a
fraction of n; for example, the proportion of active organisms to
the total number of organisms in the ith class interval.

b1 is the point of inflection of the curve.

b2  is related directly to the width of the interval.
}
\examples{
\dontrun{
x <- seq(10,30,by=2.5)
b1 <- 38
b2 <- 8
prm <- c("b1"=b1,"b2"=b2)
y_hat <- HiM(prm, x)
}

}
