% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lperu.R
\docType{data}
\name{Lperu {simodiff}}
\alias{Lperu {simodiff}}
\alias{Lperu}
\title{\emph{Lutjanus peru} mature females}
\format{
A data frame with 917 rows and 3 variables:
\describe{
\item{sex}{The sex of the organism, Female, Male or Combined ("Female")}
\item{Length}{length of fish in cm (6--65)}
\item{Binomial}{reproductive condition (0--1)}
}

A data frame with 917 rows and 2 variables
}
\source{
\url{https://github.com/Macrurido/simodiff/blob/master/data/Lperu.rda}
}
\usage{
data(Lperu)
}
\description{
The dataset contains the length and reproductive condition of female red snapper
(\emph{Lutjanus peru}) collected from the Eastern Central Pacific. The reproductive
condition is categorized as follows: 0 indicates an immature or inactive mature
organism, while 1 denotes an actively mature organism. The variables are as follows:
}
\keyword{datasets}
