% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllb.R
\name{nllb}
\alias{nllb}
\title{nllb}
\usage{
nllb(prm, df, model)
}
\arguments{
\item{prm}{A vector that contains the parameter values b1 and b2 of the selected model.}

\item{df}{A data frame, it should be organized as follows:
class mark (MC) in column 1; total number of the sample (ni)
in column 2; and number of mature organisms (mi) in column 3.}

\item{model}{A function that uses the selected model calculates the cumulative density function.}
}
\value{
The binomial negative log-likelihood value.
}
\description{
The function calculates the negative log-likelihood for a binomial
distribution of residuals.
}
\details{
The log-likelihood function, denoted as nllb, computes the fitted value y_hat
for the ith selected model and returns the binomial negative log-likelihood value.

The models included here had the same statistical assumption regarding the errors.
Each model was fit to the data, minimizing the negative value of the log-likelihood
function by using a binomial density probabilistic function (Aguirre-Villaseñor
et al., 2022).
}
\examples{
\dontrun{
x <- seq(11.25,78.75,by=2.5)
ni <- c(0,0,33,46,140,155,128,57,50,31,53,38,16,10,23,12,42,23,38,14,6,0,1,0,0,0,0,0)
mi <- c(0,0,2,0,1,2,10,2,2,2,31,16,11,8,23,10,42,23,38,14,6,0,1,0,0,0,0,0)
fo <- mi/ni
data <- cbind(x,ni,mi,fo)
data <- as.data.frame(data)

models <- list(GoM=simodiff::GoM, LoM=simodiff::LoM, GaM=simodiff::GaM)
prm <- list(GoM=c("b1"=35,"b2"=11), LoM=c("b1"=38,"b2"=4), GaM=c("b1"=26,"b2"=1.4))

out <- rep(NA,length(models))

for(i in 1:length(models)){
 model <- unlist(models[[i]])
 betas <- unlist(prm[[i]])
 out[i] <- nllb(prm=betas, df=data, model)
}
out
}

}
\seealso{
freq_mature() from the Repbio package
}
