% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(df, B, n, Imin, Imax, bin)
}
\arguments{
\item{df}{A data frame showing the length and binomial condition indicating
whether organisms are inactive (0) or actively reproducing (1).}

\item{B}{An integer that determines the number of resamplings.}

\item{n}{An integer that indicates the number of rows to be included in the
resampled data frame.}

\item{Imin}{A numeric value of the shortest organism.}

\item{Imax}{A numeric value of the longest organism.}

\item{bin}{A numeric value that represents the amplitude of the class interval.}
}
\value{
The function outputs a list where the first item contains the original
data, and items 2 to B+1 contain the resampled data frames, with B
representing the number of resamplings.

@seealso freq_mature() from simodiff package
}
\description{
This function resamples the organisms in the data frame \code{df}, which includes
the \code{Length} and \code{Binomial} columns, indicating whether each organism is
inactive (0) or active in reproduction (1).
}
\details{
The frequency of mature organisms to the total number of organisms is calculated
for each class mark using the \code{simodiff::freq_mature} function.

The results are stored in a list, with the first item representing the
original data frequencies. The variable \code{B} defines the number of
resamplings to be performed.
}
\examples{
\dontrun{
set.seed(16)

df <- Lperu
df <- df[,-1] # Sex column removed

Imin <- (trunc(min(df$Length)/5))*5
Imax <- (ceiling(max(df$Length)/5))*5
bin <- 5

B <- 3
n <- dim(df)[1]

f_resample <- resample(df, B, n, Imin, Imax, bin)
f_resample
}

}
