% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{fit_models}
\usage{
fit_models(models, prm, lista, B)
}
\arguments{
\item{models}{A list that stores the functions of selected models for fitting a   cumulative density function.}

\item{prm}{A list that contains the parameter values b1 and b2 of the selected model.}

\item{lista}{A list where the first item contains the original data, and items 2 to B+1 contain the resampled data frames.}

\item{B}{An integer that determines the number of resamplings.}
}
\value{
The function returns a list containing the stored summaries.
}
\description{
Fitting model for every database
}
\details{
All sigmoid models are fitted using the \code{optim()} function for each data set.
Each model is evaluated with the \code{nllb()} function, which calculates the
negative log-likelihood for the residuals under the assumption of a binomial
distribution.

For more details about the models, see the information about the functions
included in the simodiff package: GoM(), HiM(), LoM(), WeM() and GaM().

Within a function, an empty list named 'l_models' is created to hold the
summaries of each fit.
}
\examples{
\dontrun{
set.seed(16)
df <- load("data/Lperu.rda")
df <- df[-1]
df <- Lperu

Imin <- (trunc(min(df$Length)/5))*5
Imax <- (ceiling(max(df$Length)/5))*5
bin <- 5

B <- 3
n <- dim(df)[1]

f_resample <- simodiff::resample(df, B, n, Imin, Imax, bin)
f_resample

models <- list(GoM=simodiff::GoM, HiM=simodiff::HiM,
               LoM=simodiff::LoM, WeM=simodiff::WeM,
               GaM=simodiff::GaM)

prm <- list(GoM=c("b1"=38,"b2"=0.1), HiM=c("b1"=38,"b2"=10),
            LoM=c("b1"=38,"b2"=5), WeM=c("b1"=38,"b2"=6),
            GaM=c("b1"=21,"b2"=1.8))

fits <- fit_models(models, prm, lista=f_resample, B)
fits
}

}
