% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad.R
\name{glad}
\alias{glad}
\title{glad {simodiff}}
\usage{
glad(models, prm, df, x_label, y_label)
}
\arguments{
\item{models}{A list that contains the functions of the models.}

\item{prm}{A list that contains the parameters values for each model.}

\item{df}{A data frame with the x (class mark), y (proportion) and an empty column y_hat filled with NA}

\item{x_label}{A character vector with the label for the X-axis.}

\item{y_label}{A character vector with the label for the y-axis.}
}
\value{
a list with the new parameter values.
}
\description{
The \code{glad()} function stores the graphically determined values of the start
parameters of the models you want to analyze. Each of the models is analyzed
through a for loop; for this, the \code{simodiff::again()} function is nested.
}
\details{
For more information, please refer to the help section of the \code{simodiff::again()}
function.
}
\examples{
\dontrun{
i <- 1
models <- list(LoM= simodiff::LoM)
prm <- list(c("b1"=38,"b2"=8))
x <- seq(16.25,66.25, by=2.5)
y <- c(0.06, 0.00, 0.01, 0.01, 0.08, 0.04, 0.04,
      0.06, 0.58, 0.42, 0.69, 0.80, 1.00, 0.83,
      1.00, 1.00, 1.00, 1.00, 1.00,  NaN, 1.00)
df <- cbind(x, y)
df <- as.data.frame(df)
x_label <- "Class mark (cm)"
y_label <- "Proportion"
prm <- glad(models, prm, df, x_label, y_label)
}

}
\seealso{
again function from simodiff package.

splot function from simodiff package.
}
