% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad.R
\name{glad}
\alias{glad}
\title{glad}
\usage{
glad(models, prm, df, x_label, y_label)
}
\arguments{
\item{models}{A list that contains the functions of the models.}

\item{prm}{A list that contains the parameters values for each model.}

\item{df}{A data frame with the x (class mark), y (proportion) and an empty column y_hat filled with NA}

\item{x_label}{A character vector with the label for the X-axis.}

\item{y_label}{A character vector with the label for the y-axis.}
}
\value{
a list with the new parameter values.
}
\description{
The \code{glad()} function is a nested function that stores the values of the start
parameters graphically determined for the models you want to analyze. The
\verb{glad in()} function is encapsulated within the main \code{glad()}.
}
\details{
For more information, please refer to the help section of the \code{simodiff::glad_in()}
function.

For more details about the models, see the information about the functions
included in the simodiff package: GoM(), HiM(), LoM(), WeM() and GaM().
}
\examples{
\dontrun{
i <- 1

models <- list(GoM=simodiff::GoM, HiM=simodiff::HiM,
               LoM=simodiff::LoM, WeM=simodiff::WeM,
               GaM=simodiff::GaM)

prm <- list(GoM=c("b1"=38,"b2"=0.1), HiM=c("b1"=38,"b2"=10),
            LoM=c("b1"=38,"b2"=5), WeM=c("b1"=38,"b2"=6),
            GaM=c("b1"=21,"b2"=1.8))

x <- seq(16.25,66.25, by=2.5)
y <- c(0.06, 0.00, 0.01, 0.01, 0.08, 0.04, 0.04,
       0.06, 0.58, 0.42, 0.69, 0.80, 1.00, 0.83,
       1.00, 1.00, 1.00, 1.00, 1.00,  NaN, 1.00)

df <- cbind(x, y)
df <- as.data.frame(df)

x_label <- "Class mark (cm)"
y_label <- "Proportion"

prm <- glad(models, prm, df, x_label, y_label)

}

}
\seealso{
glad_in function from simodiff package.

splot function from simodiff package.
}
